--- docs/source/api/cuml.experimental.rst
@@ -1,22 +1,0 @@
-cuml.experimental
-=================
-
-.. warning:: The ``cuml.experimental`` module contains features that are still
-   under development. It is not recommended to depend on features in this
-   module as they may change in future releases.
-
-.. note:: Due to the nature of this module, it is not imported by default by
-   the root ``cuml`` package. Each ``experimental`` submodule must be imported
-   separately.
-
-Linear Models
--------------
-
-.. currentmodule:: cuml.experimental.linear_model
-
-.. autosummary::
-   :nosignatures:
-   :toctree: generated/
-   :template: base.rst
-
-   Lars
