+++ docs/source/api/cuml.linear_model.rst
@@ -17,3 +17,4 @@ cuml.linear_model
    ElasticNet
    MBSGDClassifier
    MBSGDRegressor
+   Lars
