+++ python/cuml/CMakeLists.txt
@@ -109,7 +109,5 @@ add_subdirectory(cuml/solvers)
 add_subdirectory(cuml/svm)
 add_subdirectory(cuml/tsa)
 
-add_subdirectory(cuml/experimental/linear_model)
-
 # Include the pth file in the built wheel
 install(FILES "_zz_cuml_accel.pth" DESTINATION ".")
