+++ python/cuml/cuml/__init__.py
@@ -46,6 +46,7 @@ from cuml.internals.global_settings import (
 from cuml.internals.outputs import set_global_output_type, using_output_type
 from cuml.kernel_ridge.kernel_ridge import KernelRidge
 from cuml.linear_model.elastic_net import ElasticNet
+from cuml.linear_model.lars import Lars
 from cuml.linear_model.lasso import Lasso
 from cuml.linear_model.linear_regression import LinearRegression
 from cuml.linear_model.logistic_regression import LogisticRegression
