--- python/cuml/cuml/experimental/linear_model/CMakeLists.txt
@@ -1,15 +1,0 @@
-# =============================================================================
-# cmake-format: off
-# SPDX-FileCopyrightText: Copyright (c) 2022-2025, NVIDIA CORPORATION.
-# SPDX-License-Identifier: Apache-2.0
-# cmake-format: on
-# =============================================================================
-
-set(cython_sources "")
-add_module_gpu_default("lars.pyx" ${lars_algo} ${linear_model_algo})
-
-rapids_cython_create_modules(
-  CXX
-  SOURCE_FILES "${cython_sources}"
-  LINKED_LIBRARIES "${cuml_sg_libraries}" MODULE_PREFIX experimental_
-)
