+++ python/cuml/tests/test_lars.py
@@ -2,8 +2,6 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-import sys
-
 import cupy as cp
 import numpy as np
 import pytest
@@ -11,7 +9,7 @@ import sklearn
 from sklearn.datasets import make_regression
 from sklearn.linear_model import Lars as skLars
 
-from cuml.experimental.linear_model import Lars as cuLars
+from cuml.linear_model import Lars as cuLars
 from cuml.testing.datasets import make_regression_dataset
 from cuml.testing.utils import (
     array_equal,
@@ -19,9 +17,6 @@ from cuml.testing.utils import (
     stress_param,
     unit_param,
 )
-
-# As tests directory is not a module, we need to add it to the path
-sys.path.insert(0, ".")
 
 
 def normalize_data(X, y):
@@ -32,6 +27,22 @@ def normalize_data(X, y):
     x_scale[x_scale == 0] = 1
     X = (X - x_mean) / x_scale
     return X, y, x_mean, x_scale, y_mean
+
+
+def test_experimental_namespace_deprecation():
+    with pytest.warns(
+        FutureWarning, match="`cuml.experimental.linear_model.Lars`"
+    ):
+        from cuml.experimental.linear_model import Lars
+
+        assert Lars is cuLars
+
+    import cuml.experimental.linear_model as ex_lm
+
+    with pytest.warns(
+        FutureWarning, match="`cuml.experimental.linear_model.Lars`"
+    ):
+        assert ex_lm.Lars is cuLars
 
 
 @pytest.mark.parametrize("datatype", [np.float32, np.float64])
